+++ lib/helpers/fields/layer-fields-builder.js
@@ -1,9 +1,9 @@
-const FieldsBuilder = require('./fields-builder')
+const QueryFieldsBuilder = require('./query-fields-builder')
 
-class LayerFieldsBuilder extends FieldsBuilder {
+class LayerFieldsBuilder extends QueryFieldsBuilder {
   static create (data, options) {
     const builder = new LayerFieldsBuilder(data, options)
-    return builder.filterByOutfields().build()
+    return builder.build()
   }
 
   build () {
