+++ lib/helpers/fields/statistics-fields-builder.js
@@ -1,5 +1,5 @@
 const _ = require('lodash')
-const FieldsBuilder = require('./fields-builder')
+const QueryFieldsBuilder = require('./query-fields-builder')
 const {
   StatisticField,
   StatisticDateField,
@@ -7,7 +7,7 @@ const {
   FieldFromKeyValue
 } = require('./field-classes')
 
-class StatisticsFieldsBuilder extends FieldsBuilder {
+class StatisticsFieldsBuilder extends QueryFieldsBuilder {
   static create (data, options) {
     const {
       metadata: {
@@ -16,20 +16,22 @@ class StatisticsFieldsBuilder extends FieldsBuilder {
       ...rest
     } = data
     const builder = new StatisticsFieldsBuilder(rest, { ...options, fields })
-    return builder.filterByOutfields().build()
+    return builder.build()
   }
 
   constructor (data, options = {}) {
     super(data, options)
     this.statisticFieldNames = getStatisticFieldsName(options.outStatistics)
-    this.statisticDateFieldRegexs = getDateFieldRegexs(this.fieldDefinitions, options.outStatistics)
+    this.dateFieldRegexs = getDateFieldRegexs(this.fieldDefinitions, options.outStatistics)
   }
 
   build () {
     const fieldDefinitions = this.fieldDefinitions || []
-    return Object.entries(this.propertiesSample)
+
+    return Object
+      .entries(this.propertiesSample)
       .map(([key, value]) => {
-        if (this._isDateField(key)) {
+        if (isDateField(this.dateFieldRegexs, key)) {
           return new StatisticDateField(key)
         }
 
@@ -46,12 +48,12 @@ class StatisticsFieldsBuilder extends FieldsBuilder {
         return new FieldFromKeyValue(key, value)
       })
   }
+}
 
-  _isDateField (fieldName) {
-    return this.statisticDateFieldRegexs.some(regex => {
-      return regex.test(fieldName)
-    })
-  }
+function isDateField (regexs, fieldName) {
+  return regexs.some(regex => {
+    return regex.test(fieldName)
+  })
 }
 
 function getDateFieldRegexs (fieldDefinitions = [], outStatistics = []) {
