+++ lib/utils.js
@@ -1,16 +1,7 @@
-const _ = require('lodash')
-const esriExtent = require('esri-extent')
 const moment = require('moment')
 const DATE_FORMATS = [moment.ISO_8601]
 
-module.exports = { getExtent, detectType, esriTypeMap }
-
-function getExtent (geojson) {
-  if (_.has(geojson, 'metadata.extent')) {
-    return geojson.metadata.extent
-  }
-  return esriExtent(geojson)
-}
+module.exports = { detectType }
 
 /**
  * returns data type based on type of value passed
@@ -27,40 +18,5 @@ function detectType (value) {
     return 'Date'
   } else {
     return 'String'
-  }
-}
-
-/**
- * returns converts type to ESRI field type
- * @param {string} type string representation of data type
- * @return {string} string representation of ESRI data type
- */
-function esriTypeMap (type) {
-  switch (type.toLowerCase()) {
-    case 'double':
-      return 'esriFieldTypeDouble'
-    case 'integer':
-      return 'esriFieldTypeInteger'
-    case 'date':
-      return 'esriFieldTypeDate'
-    case 'blob':
-      return 'esriFieldTypeBlob'
-    case 'geometry':
-      return 'esriFieldTypeGeometry'
-    case 'globalid':
-      return 'esriFieldTypeGlobalID'
-    case 'guid':
-      return 'esriFieldTypeGUID'
-    case 'raster':
-      return 'esriFieldTypeRaster'
-    case 'single':
-      return 'esriFieldTypeSingle'
-    case 'smallinteger':
-      return 'esriFieldTypeSmallInteger'
-    case 'xml':
-      return 'esriFieldTypeXML'
-    case 'string':
-    default:
-      return 'esriFieldTypeString'
   }
 }
